Move test datasets to tests/artifacts/testdata instead of tests/data
