+++ tests/acceleration/test_acceleration_framework.py
@@ -54,13 +54,13 @@ from tuning.config.acceleration_configs.quantized_lora_config import (
 from tuning.utils.import_utils import is_fms_accelerate_available
 
 # for some reason the CI will raise an import error if we try to import
-# these from tests.data
+# these from tests.artifacts.testdata
 TWITTER_COMPLAINTS_JSON_FORMAT = os.path.join(
-    os.path.dirname(__file__), "../data/twitter_complaints_json.json"
+    os.path.dirname(__file__), "../artifacts/testdata/twitter_complaints_json.json"
 )
 TWITTER_COMPLAINTS_TOKENIZED = os.path.join(
     os.path.dirname(__file__),
-    "../data/twitter_complaints_tokenized_with_maykeye_tinyllama_v0.json",
+    "../artifacts/testdata/twitter_complaints_tokenized_with_maykeye_tinyllama_v0.json",
 )
 
 # pylint: disable=import-error
