+++ tests/artifacts/testdata/trainercontroller/epoch-level-eval-loss-patience.yaml
