+++ tests/artifacts/testdata/trainercontroller/epoch-level-training-loss.yaml
