+++ tests/artifacts/testdata/trainercontroller/exposed_metrics.yaml
