+++ tests/artifacts/testdata/trainercontroller/log_controller.yaml
