+++ tests/artifacts/testdata/trainercontroller/loss_custom_metric.yaml
