+++ tests/artifacts/testdata/trainercontroller/loss_custom_operation.yaml
