+++ tests/artifacts/testdata/trainercontroller/loss_custom_operation_invalid_action.yaml
