+++ tests/artifacts/testdata/trainercontroller/loss_invalid_metric.yaml
