+++ tests/artifacts/testdata/trainercontroller/loss_invalid_operation.yaml
