+++ tests/artifacts/testdata/trainercontroller/loss_invalid_operation_action.yaml
