+++ tests/artifacts/testdata/trainercontroller/loss_invalid_trigger.yaml
