+++ tests/artifacts/testdata/trainercontroller/loss_on_threshold.yaml
