+++ tests/artifacts/testdata/trainercontroller/loss_on_threshold_with_trainer_state.yaml
