+++ tests/artifacts/testdata/trainercontroller/loss_unavailable_metric.yaml
