+++ tests/artifacts/testdata/trainercontroller/non-decreasing-training-loss.yaml
