+++ tests/artifacts/testdata/trainercontroller/thresholded-training-loss.yaml
