+++ tests/utils/test_tokenizer_data_utils.py
@@ -3,7 +3,7 @@
 from transformers import AutoModelForCausalLM, AutoTokenizer
 
 # First Party
-from tests.data import MODEL_NAME
+from tests.artifacts.testdata import MODEL_NAME
 
 # Local
 # First party
